+++ BranchUnityTestBed/Assets/Plugins/Branch/iOS/BranchiOSWrapper.mm
@@ -637,3 +637,20 @@ void _shareLinkWithLinkProperties(char *universalObjectJson, char *linkPropertie
 void _validate() {
     [[Branch getInstance] validateSDKIntegration];
 }
+
+#pragma mark - Google On Device Measurement convenience methods (ODM)
+
+void _setSDKWaitTimeForThirdPartyAPIs(double waitTime) {
+    [Branch setSDKWaitTimeForThirdPartyAPIs:waitTime];
+}
+
+void _setODMInfo(char *odmInfo, double firstOpenTimestamp) {
+    NSString *nsOdmInfo = odmInfo ? [NSString stringWithUTF8String:odmInfo] : nil;
+    NSDate *firstOpenDate = [NSDate dateWithTimeIntervalSince1970:firstOpenTimestamp];
+    [Branch setODMInfo:nsOdmInfo andFirstOpenTimestamp:firstOpenDate]; 
+}
+
+void _setAnonID(char *anonID) {
+    NSString *nsAnonID = anonID ? [NSString stringWithUTF8String:anonID] : nil;
+    [Branch setAnonID:nsAnonID];
+}